There were two test_export_integration.py tests.
Merged together the task tests and moved the other integration from tasks/ into
backend/jvm/tasks.
